Adressed code review feedback.
