+++ opcpublisher.tests/ConfigurationViaIotMethodsUnitTests .cs
@@ -1,17 +1,22 @@
+// ------------------------------------------------------------
+//  Copyright (c) Microsoft Corporation.  All rights reserved.
+//  Licensed under the MIT License (MIT). See License.txt in the repo root for license information.
+// ------------------------------------------------------------
+
 using System.Collections.Generic;
 using Xunit;
+using Microsoft.Azure.Devices.Client;
+using Newtonsoft.Json;
+using OpcPublisher.Configurations;
+using System;
+using System.IO;
+using System.Threading.Tasks;
+using Xunit.Abstractions;
+using static OpcPublisher.OpcApplicationConfiguration;
+using static OpcPublisher.Program;
 
 namespace OpcPublisher
 {
-    using Microsoft.Azure.Devices.Client;
-    using Newtonsoft.Json;
-    using System;
-    using System.IO;
-    using System.Threading.Tasks;
-    using Xunit.Abstractions;
-    using static OpcApplicationConfiguration;
-    using static Program;
-
     [Collection("Need PLC and publisher config")]
     public sealed class ConfigurationViaIotMethodUnitTests : IDisposable
     {
