+++ opcpublisher.tests/XunitCollectionDefinitions.cs
@@ -1,3 +1,8 @@
+// ------------------------------------------------------------
+//  Copyright (c) Microsoft Corporation.  All rights reserved.
+//  Licensed under the MIT License (MIT). See License.txt in the repo root for license information.
+// ------------------------------------------------------------
+
 namespace OpcPublisher
 {
     using Xunit;
