+++ opcpublisher/Configurations/PublisherTelemetryConfiguration.cs
@@ -8,395 +8,11 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using static OpcPublisher.Program;
 
-namespace OpcPublisher
+namespace OpcPublisher.Configurations
 {
-    /// <summary>
-    /// Class to control the telemetry publish, name and pattern properties.
-    /// </summary>
-    public class TelemetrySettings
-    {
-        /// <summary>
-        /// Flag to control if the value should be published.
-        /// </summary>
-        public bool? Publish
-        {
-            get => _publish;
-            set
-            {
-                if (value != null)
-                {
-                    _publish = value;
-                }
-            }
-        }
-
-        /// <summary>
-        /// The name under which the telemetry value should be published.
-        /// </summary>
-        public string Name
-        {
-            get => _name;
-            set
-            {
-                if (!string.IsNullOrEmpty(value))
-                {
-                    _name = value;
-                }
-            }
-        }
-
-        /// <summary>
-        /// The pattern which should be applied to the telemetry value.
-        /// </summary>
-        public string Pattern
-        {
-            get => _pattern;
-            set
-            {
-                if (!string.IsNullOrEmpty(value))
-                {
-                    // validate pattern
-                    try
-                    {
-                        _patternRegex = new Regex(value);
-                        _pattern = value;
-                    }
-                    catch
-                    {
-                        Logger.Fatal($"The regular expression '{value}' used for the property 'Pattern' is not a valid regular expression. Please change.");
-                        throw new Exception($"The regular expression '{value}' used for the property 'Pattern' is not a valid regular expression. Please change.");
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        /// Ctor for telemetry settings object.
-        /// </summary>
-        public TelemetrySettings()
-        {
-            _publish = null;
-            _name = null;
-            _pattern = null;
-            _patternRegex = null;
-        }
-
-        /// <summary>
-        /// Method to apply the regex to the given value if one is defined, otherwise we return the string passed in.
-        /// </summary>
-        public string PatternMatch(string stringToParse)
-        {
-            // no pattern set, return full string
-            if (_patternRegex == null)
-            {
-                return stringToParse;
-            }
-
-            // build the result string based on the pattern
-            string result = string.Empty;
-            Match match = _patternRegex.Match(stringToParse);
-            if (match.Groups[0].Success)
-            {
-                foreach (var group in match.Groups.Cast<Group>().Skip(1))
-                {
-                    result += group.Value;
-                }
-            }
-            return result;
-        }
-
-        private bool? _publish;
-        private string _name;
-        private string _pattern;
-        private Regex _patternRegex;
-    }
-
-
-    /// <summary>
-    /// Class to define the MonitoredItem related telemetry configuration.
-    /// </summary>
-    public class MonitoredItemTelemetryConfiguration
-    {
-        /// <summary>
-        /// Controls if the MonitoredItem object should be flattened.
-        /// </summary>
-        public bool? Flat
-        {
-            get => _flat;
-            set
-            {
-                if (value != null)
-                {
-                    _flat = value;
-                }
-            }
-        }
-
-        /// <summary>
-        /// The ApplicationUri value telemetry configuration.
-        /// </summary>
-        public TelemetrySettings ApplicationUri
-        {
-            get => _applicationUri;
-            set
-            {
-                _applicationUri.Publish = value.Publish;
-                _applicationUri.Name = value.Name;
-                _applicationUri.Pattern = value.Pattern;
-            }
-        }
-
-        /// <summary>
-        /// The DisplayName value telemetry configuration.
-        /// </summary>
-        public TelemetrySettings DisplayName
-        {
-            get => _displayName;
-            set
-            {
-                _displayName.Publish = value.Publish;
-                _displayName.Name = value.Name;
-                _displayName.Pattern = value.Pattern;
-            }
-        }
-
-        /// <summary>
-        /// Ctor of the object.
-        /// </summary>
-        public MonitoredItemTelemetryConfiguration()
-        {
-            _flat = null;
-            _applicationUri = new TelemetrySettings();
-            _displayName = new TelemetrySettings();
-        }
-
-        private bool? _flat;
-        private readonly TelemetrySettings _applicationUri;
-        private readonly TelemetrySettings _displayName;
-    }
-
-    /// <summary>
-    /// Class to define the Value related telemetry configuration.
-    /// </summary>
-    public class ValueTelemetryConfiguration
-    {
-        /// <summary>
-        /// Controls if the Value object should be flattened.
-        /// </summary>
-        public bool? Flat
-        {
-            get => _flat;
-            set
-            {
-                if (value != null)
-                {
-                    _flat = value;
-                }
-            }
-        }
-
-        /// <summary>
-        /// The Value value telemetry configuration.
-        /// </summary>
-        public TelemetrySettings Value
-        {
-            get => _value;
-            set
-            {
-                _value.Publish = value.Publish;
-                _value.Name = value.Name;
-                _value.Pattern = value.Pattern;
-            }
-        }
-
-        /// <summary>
-        /// The SourceTimestamp value telemetry configuration.
-        /// </summary>
-        public TelemetrySettings SourceTimestamp
-        {
-            get => _sourceTimestamp;
-            set
-            {
-                _sourceTimestamp.Publish = value.Publish;
-                _sourceTimestamp.Name = value.Name;
-                _sourceTimestamp.Pattern = value.Pattern;
-            }
-        }
-
-        /// <summary>
-        /// The StatusCode value telemetry configuration.
-        /// </summary>
-        public TelemetrySettings StatusCode
-        {
-            get => _statusCode;
-            set
-            {
-                _statusCode.Publish = value.Publish;
-                _statusCode.Name = value.Name;
-                _statusCode.Pattern = value.Pattern;
-            }
-        }
-
-        /// <summary>
-        /// The Status value telemetry configuration.
-        /// </summary>
-        public TelemetrySettings Status
-        {
-            get => _status;
-            set
-            {
-                _status.Publish = value.Publish;
-                _status.Name = value.Name;
-                _status.Pattern = value.Pattern;
-            }
-        }
-
-        /// <summary>
-        /// Ctor of the object.
-        /// </summary>
-        public ValueTelemetryConfiguration()
-        {
-            _flat = null;
-            _value = new TelemetrySettings();
-            _sourceTimestamp = new TelemetrySettings();
-            _statusCode = new TelemetrySettings();
-            _status = new TelemetrySettings();
-        }
-
-        private bool? _flat;
-        private readonly TelemetrySettings _value;
-        private readonly TelemetrySettings _sourceTimestamp;
-        private readonly TelemetrySettings _statusCode;
-        private readonly TelemetrySettings _status;
-    }
-
-    /// <summary>
-    /// Class to define the model for the telemetry configuration of an endpoint in the configuration file.
-    /// </summary>
-    public class EndpointTelemetryConfigurationModel
-    {
-        /// <summary>
-        /// Specifies the endpoint URL the telemetry should be configured for.
-        /// </summary>
-        public string ForEndpointUrl { get; set; }
-
-        /// <summary>
-        /// Specifies the configuration for the value EndpointUrl.
-        /// </summary>
-        public TelemetrySettings EndpointUrl
-        {
-            get => _endpointUrl;
-            set
-            {
-                _endpointUrl.Publish = value.Publish;
-                _endpointUrl.Name = value.Name;
-                _endpointUrl.Pattern = value.Pattern;
-            }
-        }
-
-        /// <summary>
-        /// Specifies the configuration for the value NodeId.
-        /// </summary>
-        public TelemetrySettings NodeId
-        {
-            get => _nodeId;
-            set
-            {
-                _nodeId.Publish = value.Publish;
-                _nodeId.Name = value.Name;
-                _nodeId.Pattern = value.Pattern;
-            }
-        }
-
-        public TelemetrySettings ExpandedNodeId
-        {
-            get => _expandedNodeId;
-            set
-            {
-                _expandedNodeId.Publish = value.Publish;
-                _expandedNodeId.Name = value.Name;
-                _expandedNodeId.Pattern = value.Pattern;
-            }
-        }
-
-        /// <summary>
-        /// Specifies the configuration for the value MonitoredItem.
-        /// </summary>
-        public MonitoredItemTelemetryConfiguration MonitoredItem
-        {
-            get => _monitoredItem;
-            set
-            {
-                _monitoredItem.Flat = value.Flat;
-                _monitoredItem.ApplicationUri = value.ApplicationUri;
-                _monitoredItem.DisplayName = value.DisplayName;
-            }
-        }
-
-        /// <summary>
-        /// Specifies the configuration for the value Value.
-        /// </summary>
-        public ValueTelemetryConfiguration Value
-        {
-            get => _value;
-            set
-            {
-                _value.Flat = value.Flat;
-                _value.Value = value.Value;
-                _value.SourceTimestamp = value.SourceTimestamp;
-                _value.StatusCode = value.StatusCode;
-                _value.Status = value.Status;
-            }
-        }
-
-        /// <summary>
-        /// Ctor for an object.
-        /// </summary>
-        public EndpointTelemetryConfigurationModel()
-        {
-            ForEndpointUrl = null;
-            _endpointUrl = new TelemetrySettings();
-            _nodeId = new TelemetrySettings();
-            _expandedNodeId = new TelemetrySettings();
-            _monitoredItem = new MonitoredItemTelemetryConfiguration();
-            _value = new ValueTelemetryConfiguration();
-        }
-
-        private readonly TelemetrySettings _endpointUrl;
-        private readonly MonitoredItemTelemetryConfiguration _monitoredItem;
-        private readonly ValueTelemetryConfiguration _value;
-        private readonly TelemetrySettings _nodeId;
-        private readonly TelemetrySettings _expandedNodeId;
-    }
-
-    /// <summary>
-    /// Class to define the telemetryconfiguration.json configuration file layout.
-    /// </summary>
-    public class TelemetryConfigurationFileModel
-    {
-        /// <summary>
-        /// Default settings for all endpoints without specific configuration.
-        /// </summary>
-        public EndpointTelemetryConfigurationModel Defaults { get; set; }
-
-        /// <summary>
-        /// Endpoint specific configuration.
-        /// </summary>
-        public List<EndpointTelemetryConfigurationModel> EndpointSpecific { get; }
-
-        /// <summary>
-        /// Ctor for the telemetry configuration.
-        /// </summary>
-        public TelemetryConfigurationFileModel()
-        {
-            EndpointSpecific = new List<EndpointTelemetryConfigurationModel>();
-        }
-    }
-
     public class PublisherTelemetryConfiguration
     {
         public const string EndpointUrlNameDefault = "EndpointUrl";
