+++ opcpublisher/HubClient.cs
@@ -4,6 +4,7 @@
 // ------------------------------------------------------------
 
 using Microsoft.Azure.Devices.Client;
+using OpcPublisher.Interfaces;
 using System;
 using System.Threading.Tasks;
