+++ opcpublisher/Interfaces/IPublisherNodeConfiguration.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace OpcPublisher
+namespace OpcPublisher.Interfaces
 {
     public interface IPublisherNodeConfiguration
     {
