+++ opcpublisher/IotEdgeHubCommunication.cs
@@ -3,6 +3,7 @@
 //  Licensed under the MIT License (MIT). See License.txt in the repo root for license information.
 // ------------------------------------------------------------
 
+using OpcPublisher.Interfaces;
 using System;
 using static OpcPublisher.Program;
